refactor: Swiper
